Move caps to bottom key, map ctrl space
Caps is used in combination with a bunch of other keys, but control
space is just hard and annoying to hit when control is next to space, so
let's try giving it a dedicated key! Worst case we could make it an edit
layer binding for the left hand or something like that.